+++ terraform/terraform_client.go
@@ -3,7 +3,6 @@ package terraform
 
 import (
 	"fmt"
-	"os"
 	"os/exec"
 	"regexp"
 
@@ -52,8 +51,8 @@ func (c *Client) Version() *version.Version {
 }
 
 // RunCommandWithVersion executes the provided version of terraform with
-// the provided args and envVars in path.
-func (c *Client) RunCommandWithVersion(log *logging.SimpleLogger, path string, args []string, envVars []string, v *version.Version) (string, error) {
+// the provided args in path.
+func (c *Client) RunCommandWithVersion(log *logging.SimpleLogger, path string, args []string, v *version.Version) (string, error) {
 	tfExecutable := "terraform"
 	// if version is the same as the default, don't need to prepend the version name to the executable
 	if !v.Equal(c.defaultVersion) {
@@ -61,11 +60,6 @@ func (c *Client) RunCommandWithVersion(log *logging.SimpleLogger, path string, a
 	}
 	terraformCmd := exec.Command(tfExecutable, args...)
 	terraformCmd.Dir = path
-	if len(envVars) > 0 {
-		// append current process's environment variables
-		// this is to prevent the $PATH variable being removed from the environment
-		terraformCmd.Env = append(os.Environ(), envVars...)
-	}
 	out, err := terraformCmd.CombinedOutput()
 	commandStr := strings.Join(terraformCmd.Args, " ")
 	if err != nil {
@@ -80,21 +74,21 @@ func (c *Client) RunCommandWithVersion(log *logging.SimpleLogger, path string, a
 // RunInitAndEnv executes "terraform init" and "terraform env select" in path.
 // env is the environment to select and extraInitArgs are additional arguments
 // applied to the init command.
-func (c *Client) RunInitAndEnv(log *logging.SimpleLogger, path string, env string, extraInitArgs []string, envVars []string, version *version.Version) ([]string, error) {
+func (c *Client) RunInitAndEnv(log *logging.SimpleLogger, path string, env string, extraInitArgs []string, version *version.Version) ([]string, error) {
 	var outputs []string
 	// run terraform init
-	output, err := c.RunCommandWithVersion(log, path, append([]string{"init", "-no-color"}, extraInitArgs...), envVars, version)
+	output, err := c.RunCommandWithVersion(log, path, append([]string{"init", "-no-color"}, extraInitArgs...), version)
 	if err != nil {
 		return nil, err
 	}
 	outputs = append(outputs, output)
 
 	// run terraform env new and select
-	output, err = c.RunCommandWithVersion(log, path, []string{"env", "select", "-no-color", env}, envVars, version)
+	output, err = c.RunCommandWithVersion(log, path, []string{"env", "select", "-no-color", env}, version)
 	if err != nil {
 		// if terraform env select fails we will run terraform env new
 		// to create a new environment
-		output, err = c.RunCommandWithVersion(log, path, []string{"env", "new", "-no-color", env}, envVars, version)
+		output, err = c.RunCommandWithVersion(log, path, []string{"env", "new", "-no-color", env}, version)
 		if err != nil {
 			return nil, err
 		}