+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # v0.1.1 (Unreleased)
+### Backwards Incompatibilities / Notes:
+* `--aws-assume-role-arn` and `--aws-region` flags removed. Instead, to name the
+assume role session with the GitHub username of the user running the Atlantis command
+use the `atlantis_user` terraform variable alongside Terraform's
+[built-in support](https://www.terraform.io/docs/providers/aws/#assume-role) for assume role
+(see https://github.com/hootsuite/atlantis/blob/master/README.md#assume-role-session-names)
+
 ### Improvements
-* Support for HTTPS cloning using Github username and token provided to atlantis server ([#117](https://github.com/hootsuite/atlantis/pull/117))
-* Set custom user agent for AWS interactions by atlantis server ([#115](https://github.com/hootsuite/atlantis/pull/115))
+* Support for HTTPS cloning using GitHub username and token provided to atlantis server ([#117](https://github.com/hootsuite/atlantis/pull/117))
 * Adding `post_plan` and `post_apply` commands ([#102](https://github.com/hootsuite/atlantis/pull/102))
 * Adding the ability to verify webhook secret ([#120](https://github.com/hootsuite/atlantis/pull/120))